@@ -2,17 +2,17 @@
   "name": "chime-sdk-meeting-demo",
   "version": "0.1.0",
   "dependencies": {
-    "@types/react": "^16.14.30",
-    "@types/react-dom": "^16.9.16",
+    "@types/react": "^18.2.7",
+    "@types/react-dom": "^18.2.4",
     "@types/styled-components": "^4.4.3",
     "@types/webpack-env": "^1.17.0",
     "amazon-chime-sdk-component-library-react": "file:../../amazon-chime-sdk-component-library-react",
-    "amazon-chime-sdk-js": "^3.10.0",
+    "amazon-chime-sdk-js": "^3.14.0",
     "aws-sdk": "^2.1354.0",
     "fs-extra": "^10.1.0",
     "lodash.isequal": "^4.5.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^5.3.3",
     "styled-components": "^5.3.5",
     "styled-system": "^5.1.5",
